feat: member 테이블 변경사항 반영
- id type을 int에서 long으로 변경
- QR 이미지 경로 컬럼 추가

+++ bible/src/main/java/com/library/bible/member/dto/MemberRequestDto.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class MemberRequestDto {
-	private int memId;
+	private long memId;
 	private String memName;
 	private String memPassword;
